Move errors and schema out of utils
