+++ packages/sdk/src/config/errors.ts
