+++ packages/sdk/src/config/types.ts
@@ -1,4 +1,4 @@
 import { z } from "zod";
-import { configSchema } from "./utils/schema";
+import { configSchema } from "./schema";
 
 export type VlayerContextConfig = z.infer<typeof configSchema>;
